⚡ Correction history: `int[]` -> `short[]` (#2306)

+++ src/Lynx/Search/Helpers.cs
@@ -199,7 +199,7 @@ public sealed partial class Engine
 
         // Common update logic
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        static int UpdateCorrectionHistory(int previousCorrectedScore, int scaledBonus, int weight)
+        static short UpdateCorrectionHistory(short previousCorrectedScore, int scaledBonus, int weight)
         {
             const int weightScale = 256;
             var maxIncrement = Configuration.EngineSettings.CorrHistory_MaxRawBonus;
@@ -212,7 +212,7 @@ public sealed partial class Engine
 
             weightedEval = Math.Clamp(weightedEval, previousCorrectedScore - maxIncrement, previousCorrectedScore + maxIncrement);
 
-            return Math.Clamp(weightedEval, -maxVal, +maxVal);
+            return (short)Math.Clamp(weightedEval, -maxVal, +maxVal);
         }
     }
